pmd_editor/requirements.txt
Django==3.2.5
django-multiselectfield==1.0.1
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
django-cors-headers==3.7.0
mysqlclient==2.2.7
gunicorn==20.1.0
pytz==2025.2
requests==2.31.0
